requirements.txt
flask==3.0.0
pymilvus==2.3.4
langchain-ollama==0.1.0
langchain-core==0.1.0
langgraph==0.2.0
langchain-mcp-adapters==0.1.0
typing-extensions==4.9.0
python-dotenv==1.0.0
pydantic==2.5.0
fastmcp==0.1.0
beautifulsoup4==4.12.2
selenium==4.15.0
requests==2.31.0
nest-asyncio==1.5.8 